back-end/requirements.txt
Flask
flask-cors
tensorflow-cpu
grpcio>=1.24.3,<2.0
pillow
numpy
torch
torchvision